# application/services/custom_services/medicationmanager/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: healthmonkstart63 open api specification
   description: medicationmanager API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8022
     description: medicationmanager
tags:
- name: medication
  description: represents a medicine
paths:
   "/medication/{id}":
      delete:
         tags:
         -  medication
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  medication
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/medication/get/search":
      get:
         tags:
         -  medication
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/medication/get/update":
      put:
         tags:
         -  medication
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/medication"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/medication":
      put:
         tags:
         -  medication
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/medication"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  medication
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  medication
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/medication"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      medication:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            type:
               type: "string"
               description: Description
            dosage-milligram:
               type: "string"
               description: Description
            active:
               type: "boolean"
               description: Description
            prescription-id:
               type: "string"
               description: Description
            schedule:
               type: "string"
               description: this should key off of ID not name